Update 03-lazy-loading.md
"Amazon FSx for Lustre Lazy Load Workshop" link is broken. I tried to find the actual one but am not sure if this is the right one.

Signed-off-by: Alexandre Gobeaux gobeaa@amazon.com

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.