Remove the unnecessary jump EC2 -> Cloud9 -> EC2
Move "Launch the EC2 Dashbaoard" down right before "Create and EC2
instace". This way the logical connection is not broken and we are not
jumping around between Services.

The feedback we received was "The next step is to go to Cloud9. So,
what do I do with this one? Somehow the flow is odd".