Make the instructions work even if the user does not pay attention.
Feedback we received was "“Type q to exit the help pages” should be a
sub bullet of “the general AWS CLI help”Some folks may already be
copying and pasting these commands mechanically and do not realize
they are inside of a help window. This happened to me the first time I
went through these examples.".